+++ src/uranometria/annotate/model.py
@@ -48,15 +48,9 @@ def _image_size(image):
 
 
 def _links(designation, common=None):
-    links = {
-        "simbad": "https://simbad.cds.unistra.fr/simbad/sim-id?Ident="
-        + designation.replace(" ", "+")
-    }
-    if designation.startswith("M") and designation[1:].isdigit():
-        links["wikipedia"] = f"https://en.wikipedia.org/wiki/Messier_{designation[1:]}"
-    elif common:
-        links["wikipedia"] = "https://en.wikipedia.org/wiki/" + common.replace(" ", "_")
-    return links
+    from ..catalog import object_links
+
+    return {label.lower(): url for label, url in object_links(designation, common)}
 
 
 def _harmonize_pair_distances(objects):